Add localization to the OO-DOCI executable. Add more output.
